requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
constantly==15.1.0
cryptography==41.0.5
daphne==4.0.0
Django==4.2.6
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.3.0
redis==5.0.1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.8.0
zope.interface==6.1
whitenoise==6.2.0
djangorestframework==3.14.0
